Lectures/CMU_10-703/HW2/requirements.txt
attrs
gym
h5py
keras
matplotlib
numpy
pytest
pillow
protobuf>=3.0
pydot-ng
scipy
semver
-e .
